Moved camera permission check to dispatcher fragment
