+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/SubmissionDispatcherFragment.kt
@@ -1,16 +1,22 @@
 package de.rki.coronawarnapp.ui.submission
 
+import android.Manifest
+import android.content.pm.PackageManager
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSubmissionDispatcherBinding
 import de.rki.coronawarnapp.ui.BaseFragment
+import de.rki.coronawarnapp.util.CameraPermissionHelper
 
 class SubmissionDispatcherFragment : BaseFragment() {
 
     companion object {
+        private const val REQUEST_CAMERA_PERMISSION_CODE = 1
         private val TAG: String? = SubmissionDispatcherFragment::class.simpleName
     }
 
@@ -33,9 +39,7 @@ class SubmissionDispatcherFragment : BaseFragment() {
 
     private fun setButtonOnClickListener() {
         binding.submissionDispatcherQr.dispatcherCard.setOnClickListener {
-            doNavigate(
-                SubmissionDispatcherFragmentDirections.actionSubmissionDispatcherFragmentToRegisterQRCodeFragment()
-            )
+            checkForCameraPermission()
         }
         binding.submissionDispatcherTanCode.dispatcherCard.setOnClickListener {
             doNavigate(
@@ -45,5 +49,77 @@ class SubmissionDispatcherFragment : BaseFragment() {
         binding.submissionDispatcherTanTele.dispatcherCard.setOnClickListener {
             Log.i(TAG, "TAN tele pressed")
         }
+    }
+
+    private fun checkForCameraPermission() {
+        if (!CameraPermissionHelper.hasCameraPermission(requireActivity())) {
+            if (shouldShowRequestPermissionRationale(Manifest.permission.CAMERA)) {
+                showCameraPermissionRationaleDialog()
+            } else {
+                requestPermissions(
+                    arrayOf(Manifest.permission.CAMERA),
+                    REQUEST_CAMERA_PERMISSION_CODE
+                )
+            }
+        } else {
+            cameraPermissionIsGranted()
+        }
+    }
+
+    private fun showCameraPermissionRationaleDialog() {
+        val alertDialog: AlertDialog = requireActivity().let {
+            val builder = AlertDialog.Builder(it)
+            builder.apply {
+                setTitle(R.string.register_qr_code_scan_permission_rationale_dialog_headline)
+                setMessage(R.string.register_qr_code_scan_permission_rationale_dialog_body)
+                setPositiveButton(
+                    R.string.register_qr_code_scan_permission_rationale_dialog_button_positive
+                ) { _, _ ->
+                    requestPermissions(
+                        arrayOf(Manifest.permission.CAMERA),
+                        REQUEST_CAMERA_PERMISSION_CODE
+                    )
+                }
+                setNegativeButton(
+                    R.string.register_qr_code_scan_permission_rationale_dialog_button_negative
+                ) { _, _ -> }
+            }
+            builder.create()
+        }
+        alertDialog.show()
+    }
+
+    override fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<out String>,
+        grantResults: IntArray
+    ) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+        if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+            cameraPermissionIsGranted()
+        } else {
+            showCameraPermissionDeniedDialog()
+        }
+    }
+
+    private fun showCameraPermissionDeniedDialog() {
+        val alertDialog: AlertDialog = requireActivity().let {
+            val builder = AlertDialog.Builder(it)
+            builder.apply {
+                setTitle(R.string.register_qr_code_scan_permission_denied_dialog_headline)
+                setMessage(R.string.register_qr_code_scan_permission_denied_dialog_body)
+                setPositiveButton(
+                    R.string.register_qr_code_scan_permission_denied_dialog_button_positive
+                ) { _, _ -> }
+            }
+            builder.create()
+        }
+        alertDialog.show()
+    }
+
+    private fun cameraPermissionIsGranted() {
+        doNavigate(
+            SubmissionDispatcherFragmentDirections.actionSubmissionDispatcherFragmentToRegisterQRCodeFragment()
+        )
     }
 }
